requirements.txt
aiohttp==3.6.2
redis==3.5.2
pytest-aiohttp==0.3.0
